refactor: cleaning up annotation feature in frontend
